Move tabs screens to tabs
