+++ apps/router-e2e/__e2e__/native-navigation/__tests__/tabs/assert-all-tabs-visible.yml
