+++ apps/router-e2e/__e2e__/native-navigation/__tests__/tabs/link-preview-menu.yml
@@ -2,8 +2,8 @@ appId: com.expo.routere2e
 name: Link preview menu
 ---
 - stopApp: com.expo.routere2e
-- openLink: router-e2e://
-- runFlow: close-warning.yml
+- openLink: router-e2e://tabs
+- runFlow: ../close-warning.yml
 - assertVisible: "Index screen"
 
 - scrollUntilVisible:
