+++ apps/router-e2e/__e2e__/native-navigation/__tests__/tabs/link-preview.yml
@@ -2,8 +2,8 @@ appId: com.expo.routere2e
 name: Link preview
 ---
 - stopApp: com.expo.routere2e
-- openLink: router-e2e://faces
-- runFlow: close-warning.yml
+- openLink: router-e2e://tabs/faces
+- runFlow: ../close-warning.yml
 - assertVisible: "Face Gallery"
 - longPressOn:
     id: "link-face-0"
