+++ apps/router-e2e/__e2e__/native-navigation/app/tabs/dynamic.tsx
