+++ apps/router-e2e/__e2e__/native-navigation/app/tabs/faces/index.tsx
@@ -1,6 +1,6 @@
 import { ScrollView } from 'react-native';
 
-import { Faces } from '../../components/faces';
+import { Faces } from '../../../components/faces';
 
 export default function Index() {
   return (
